Moved Disclaimer to components
