+++ apps/client/src/components/Conclusion/index.ts
@@ -1,5 +1,4 @@
 import ConclusionOutcome from "./ConclusionOutcome";
-import Disclaimer from "./Disclaimer";
 import NeedPermitContent from "./NeedPermitContent";
 import NeedPermitFooter from "./NeedPermitFooter";
 import NewCheckerModal from "./NewCheckerModal";
@@ -7,7 +6,6 @@ import NoPermitDescription from "./NoPermitDescription";
 
 export {
   ConclusionOutcome,
-  Disclaimer,
   NeedPermitContent,
   NeedPermitFooter,
   NewCheckerModal,
