+++ apps/client/src/components/Disclaimer.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { DISCLAIMER_TEXT } from "../../utils/test-ids";
-import { cleanup, render } from "../../utils/test-utils";
+import { DISCLAIMER_TEXT } from "../utils/test-ids";
+import { cleanup, render } from "../utils/test-utils";
 import Disclaimer from "./Disclaimer";
 
 afterEach(cleanup);
